# BUG/PERF: Sort mixed-int in Py3, fix Index.difference