… doc/source/whatsnew/v0.19.0.txt
… pandas/tools/tests/test_join.py