@jreback 2016-07-03
@pijucha looking really good. just a couple of comments.

@sinhrks if you can have a look.