@pijucha 2016-07-04
I just noticed that `Index.sort_values` may benefit from `safe_sort`. So maybe I'll rewrite it too.

---
Also, I was thinking about MultiIndex #13504. I see two possible solutions:
1. make MultiIndex use `Index.difference` and `symmetric_difference`, and probably modify `MultiIndex._shallow_copy` (or `.from_tuples`)
2. just copy `Index.difference` to `MultiIndex.difference` but without the sorting block (I think it's not needed, `sortlevel` will do the job).
Solution 2 means some code duplication but it's faster.
I don't want to interfere with the other PR, so I'll probably add a separate comment there.