@jreback 2016-07-06
ok @pijucha looks good. minor doc updates. ping when green. other changes leave to other PR's