@jreback 2016-07-10
@pijucha pls rebase, merging lots of stuff. ping when green. I think this lgtm.