@pijucha 2016-07-14
@jreback I see several merges have been made in the meantime. Should I rebase again?