@pijucha 2016-07-14
@jreback Rebased, tests green.