@jreback 2016-07-14
@pijucha looks really good. just some tests splitting, minor corrections.

ping on green.