@pijucha 2016-07-15
@jreback OK, cleaned some tests. It's green.