@jreback 2016-07-15
minor doc change. ping on green.