@jreback 2016-07-15
some failures on windows.  we ultimately test on appveyor.com (you just need a free account), but its very slow. so easiest to spin up a vm if you can.

these are prob because you need to use ``_ensure_platform_int`` when you are using an indexer (on windows things like ``.take`` and ``.putmask`` take a ``int32`` rather than standard ``int64``. in this particular case you actually need to ``_ensure_int64`` (as you actually have platform ints up front).


```
(pandas3.5) C:\Users\conda\Documents\pandas3.5>nosetests pandas/tests/indexes
..........................................................................................................................................
..........................................................................................................................................
..........................................................................................................................................
.........................................................................S................................................................
......................................................................................................E...................................
................................................E...............................................
======================================================================
ERROR: test_join_non_unique (pandas.tests.indexes.test_numeric.TestInt64Index)
----------------------------------------------------------------------
Traceback (most recent call last):
  File "C:\Users\conda\Documents\pandas3.5\pandas\tests\indexes\test_numeric.py", line 755, in test_join_non_unique
    joined, lidx, ridx = left.join(left, return_indexers=True)
  File "C:\Users\conda\Documents\pandas3.5\pandas\indexes\base.py", line 2635, in join
    return_indexers=return_indexers)
  File "C:\Users\conda\Documents\pandas3.5\pandas\indexes\base.py", line 2720, in _join_non_unique
    sort=True)
  File "C:\Users\conda\Documents\pandas3.5\pandas\tools\merge.py", line 909, in _get_join_indexers
    return join_func(lkey, rkey, count, **kwargs)
  File "pandas\src\join.pyx", line 51, in pandas.algos.left_outer_join (pandas\algos.c:33321)
    def left_outer_join(ndarray[int64_t] left, ndarray[int64_t] right,
ValueError: Buffer dtype mismatch, expected 'int64_t' but got 'long'

======================================================================
ERROR: test_join_non_unique (pandas.tests.indexes.test_range.TestRangeIndex)
----------------------------------------------------------------------
Traceback (most recent call last):
  File "C:\Users\conda\Documents\pandas3.5\pandas\tests\indexes\test_range.py", line 524, in test_join_non_unique
    res, lidx, ridx = self.index.join(other, return_indexers=True)
  File "C:\Users\conda\Documents\pandas3.5\pandas\indexes\range.py", line 435, in join
    return super(RangeIndex, self).join(other, how, level, return_indexers)
  File "C:\Users\conda\Documents\pandas3.5\pandas\indexes\base.py", line 2642, in join
    return_indexers=return_indexers)
  File "C:\Users\conda\Documents\pandas3.5\pandas\indexes\base.py", line 2720, in _join_non_unique
    sort=True)
  File "C:\Users\conda\Documents\pandas3.5\pandas\tools\merge.py", line 909, in _get_join_indexers
    return join_func(lkey, rkey, count, **kwargs)
  File "pandas\src\join.pyx", line 51, in pandas.algos.left_outer_join (pandas\algos.c:33321)
    def left_outer_join(ndarray[int64_t] left, ndarray[int64_t] right,
ValueError: Buffer dtype mismatch, expected 'int64_t' but got 'long'

----------------------------------------------------------------------
Ran 796 tests in 10.849s

FAILED (SKIP=1, errors=2)
```