@pijucha 2016-07-15
OK, I think I've found the problem but need some time to test it on windows.