@pijucha 2016-07-17
@jreback The fix was straightforward and it should be ok now.

However, I'm getting these two errors on windows (64bit)  - also on the master branch, so they're independent of this PR. Either I messed up the windows setup or it's not being tested thoroughly.

```
======================================================================
FAIL: test_next (pandas.io.tests.test_common.TestMMapWrapper)
----------------------------------------------------------------------
Traceback (most recent call last):
  File "D:\GitHub\pandas-pijucha\pandas\io\tests\test_common.py", line 141, in t
est_next
    self.assertEqual(next_line, line)
AssertionError: 'a,b,c\r\n' != 'a,b,c\n'
- a,b,c
?      -
+ a,b,c

======================================================================
FAIL: test_alignment_non_pandas (pandas.tests.frame.test_operators.TestDataFrame
Operators)
----------------------------------------------------------------------
Traceback (most recent call last):
  File "D:\GitHub\pandas-pijucha\pandas\tests\frame\test_operators.py", line 120
2, in test_alignment_non_pandas
    Series([1, 2, 3], index=df.index))
  File "D:\GitHub\pandas-pijucha\pandas\util\testing.py", line 1157, in assert_s
eries_equal
    assert_attr_equal('dtype', left, right)
  File "D:\GitHub\pandas-pijucha\pandas\util\testing.py", line 882, in assert_at
tr_equal
    left_attr, right_attr)
  File "D:\GitHub\pandas-pijucha\pandas\util\testing.py", line 1021, in raise_as
sert_detail
    raise AssertionError(msg)
AssertionError: Attributes are different

Attribute "dtype" are different
[left]:  int32
[right]: int64None
```