@jreback 2016-07-17
2nd is already fixed
make sure u r rebased
first was fixed a while back as well