@pijucha 2016-07-17
Ah, then it's probably ok. I tested 0.18.1+202.g0a70b5f on windows.
(This commit is already rebased to the current master. I'll check on windows in a minute.)