@pijucha 2016-07-17
Yes, everything is all right.
The second error is fixed, and the first one was apparently due to my git `autocrlf` config.