@jreback 2016-07-19
is #13432 completely closed by this? if not pls make checkboxes in the top (and tick off the things that are done and leave open things that are not).