@pijucha 2016-07-19
@jreback Thanks.

As for #13432, the issues with `difference` and `symmetric_difference` are solved. But `union` and `intersection` are still open. I'll comment there soon.