@jreback 2016-07-19
@pijucha this is failing on windows / py2.7 (3 is clean): https://ci.appveyor.com/project/jreback/pandas-465/build/1.0.762/job/ghokkwmadwog983y

not sure what is going on